requirements.txt
django==2.2.1
djangorestframework==3.9.4
django-rest-swagger==2.2.0
